# apps/api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for the Completionist application.",
        "title": "Completionist API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/lists/tmdb/movie/{tmdb_id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Imports a movie from TMDb into the user's list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lists"
                ],
                "summary": "Add Movie from TMDb",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TMDb Movie ID",
                        "name": "tmdb_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "List details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.addFromTMDBBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.UserListItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/lists/tmdb/tv/{tmdb_id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Imports a TV show from TMDb into the user's list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lists"
                ],
                "summary": "Add TV Show from TMDb",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TMDb TV ID",
                        "name": "tmdb_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "List details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.addFromTMDBBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.UserListItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/search/movies": {
            "get": {
                "description": "Search for movies using The Movie Database API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Search Movies (TMDb)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tmdb.SearchResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/search/tv": {
            "get": {
                "description": "Search for TV shows using The Movie Database API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Search TV Shows (TMDb)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tmdb.SearchResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.addFromTMDBBody": {
            "type": "object",
            "properties": {
                "progressData": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.ItemStatus"
                }
            }
        },
...
        "models.ItemStatus": {
            "type": "string",
            "enum": [
                "planning",
                "current",
                "completed",
                "paused",
                "dropped"
            ],
            "x-enum-varnames": [
                "StatusPlanning",
                "StatusCurrent",
                "StatusCompleted",
                "StatusPaused",
                "StatusDropped"
            ]
        },
        "models.ItemType": {
            "type": "string",
            "enum": [
                "manga",
                "game",
                "book",
                "movie",
                "series",
                "music"
            ],
            "x-enum-varnames": [
                "ItemTypeManga",
                "ItemTypeGame",
                "ItemTypeBook",
                "ItemTypeMovie",
                "ItemTypeSeries",
                "ItemTypeMusic"
            ]
        },
...
        "models.MediaItem": {
            "type": "object",
            "properties": {
                "coverImageUrl": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "criticRatingCount": {
                    "type": "integer"
                },
                "criticRatingValue": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "externalId": {
                    "type": "string"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "itemType": {
                    "$ref": "#/definitions/models.ItemType"
                },
                "metadata": {
                    "type": "string"
                },
                "releaseDate": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userRatingExternal": {
                    "type": "number"
                }
            }
        },
...
        "models.UserListItem": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "itemType": {
                    "$ref": "#/definitions/models.ItemType"
                },
                "media": {
                    "$ref": "#/definitions/models.MediaItem"
                },
                "mediaItemId": {
                    "type": "string"
                },
                "progressData": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.ItemStatus"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "tmdb.SearchItem": {
            "type": "object",
            "properties": {
                "first_air_date": {
                    "type": "string"
                },
                "genre_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "media_type": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "poster_path": {
                    "type": "string"
                },
                "release_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "tmdb.SearchResult": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/tmdb.SearchItem"
                    }
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_results": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}